// src/app/enums/currenciesEnum.ts
export enum CurrenciesEnum {
  AUD = 'AUD',
  BGN = 'BGN',
  BRL = 'BRL',
  CAD = 'CAD',
  CHF = 'CHF',
  CNY = 'CNY',
  CYP = 'CYP',
  CZK = 'CZK',
  DKK = 'DKK',
  EEK = 'EEK',
  EUR = 'EUR',
  GBP = 'GBP',
  HKD = 'HKD',
  HRK = 'HRK',
  HUF = 'HUF',
  IDR = 'IDR',
  ILS = 'ILS',
  INR = 'INR',
  ISK = 'ISK',
  JPY = 'JPY',
  KRW = 'KRW',
  LTL = 'LTL',
  LVL = 'LVL',
  MTL = 'MTL',
  MXN = 'MXN',
  MYR = 'MYR',
  NOK = 'NOK',
  NZD = 'NZD',
  PHP = 'PHP',
  PLN = 'PLN',
  ROL = 'ROL',
  RON = 'RON',
  RUB = 'RUB',
  SEK = 'SEK',
  SGD = 'SGD',
  SIT = 'SIT',
  SKK = 'SKK',
  THB = 'THB',
  TRL = 'TRL',
  TRY = 'TRY',
  USD = 'USD',
  ZAR = 'ZAR',
}